[PERF] sale_stock: fix memory error in anglo_saxon price
Inside `_stock_account_get_anglo_saxon_price_unit` we are retrieving
the posted cogs linked to the account.moves. To do that, currently
we are retrieving every cogs and then calling `filtered` on the result.

This can lead to a huge number of cogs to fetch in large databases.
As this filtered call can easily be turned into a search_domain,
we replace it by a search call. That way the number of cogs retrieved
is way smaller. This has the effect of greatly reducing the memory
footprint of the method, as fetching all the fields on a large number
of cogs quickly filled up the memory prior to this commit.

closes odoo/odoo#186824

X-original-commit: 436f7ca6cd61fa8d48f8fc4cb081b445eaf5e150
Signed-off-by: William Henrotin (whe) <whe@odoo.com>
Signed-off-by: Aurélien van Delft (avd) <avd@odoo.com>